changed scaleCalibration to double

+++ src/Storage.cpp
@@ -58,7 +58,7 @@ typedef struct __attribute__((packed)) {
     char wifiPassword[25 + 1];
     double weightSetpoint;
     double steamkp;
-    float scaleCalibration;
+    double scaleCalibration;
     double scaleKnownWeight;
     double steamSetpoint;
 } sto_data_t;
